serious rewrite to support async updates
updates now occur as messages which are dispatched to the
game state.

puzzle and game state are now separate objects

most things should work, but it's not as thoroughly tested
as the previous code, so some things might be broken.

no migration from previous puzzles, as I don't think this
is getting real use yet.