client UI now has main screen
the main screen will show a list of previous puzzles you have joined
or uploaded, if you had joined one before.  If not it will show a
puzzle entry screen.

Fix bug where clicking on a cell wouldn't toggle between row and
band mode.

Disable text selection on solve screen except for clues.  This fixes
some bugs where the screen would scroll when dragging, in a way
that couldn't be undone.  It also fixes a selection cursor flashing
in the middle of the grid.

Down-side is that you can't copy the solve, but that's ok I think.

Also, change server behavior to only return puzzles that you uploaded.
Unless you're an admin, which means your private UUID is in a list
set on the server, in which case you'll get all the ones available.